move data select component
